+++ parser/config.go
@@ -142,6 +142,7 @@ var SupportedFunctions = map[string]int{
 	"isdeficient": 1,
 	"digitsum":   1,
 	"primorial": 1,
+	"lambertw":  1,
 	"digitcount": 1, // Collatz stopping time
 	"prod":      -1,
 	"min":       -1, // variadic